requirements.txt
tensorflow==2.12.0
numpy==1.24.3
matplotlib==3.7.1
tkinter
pillow==10.0.0